+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Verb Changelog
+## master
+- Request headers can now be commented out using `#`.
+- Surrounding whitespace is now automatically removed from header names and values.
+
 ## **1.2.0** - 2020-01-16
 - Request specifications can now be exported to `curl`.