+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Arguments are now separated by (escaped) newlines.
 - Added `verb-export-request-on-point-curl`, `verb-export-request-on-point-verb` and `verb-export-request-on-point-human` commands.
 - Bind <kbd>C-c C-r C-u</kbd> to `verb-export-request-on-point-curl` in Verb mode.
+- Added `verb-set-var` command.
 
 ## **1.2.0** - 2020-01-16
 - Request specifications can now be exported to `curl`.