+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## master
+## **1.3.0** - 2020-01-17
 - Request headers can now be commented out using `#`.
 - Surrounding whitespace is now automatically removed from header names and values.
 - Changed `curl` export format slightly: